Namespace root tracking working
This was a bad implementation. If namespaces aren't going to be in the heap, I need at least to track slots properly not loop over everything. I also don't love that namespaces are part of compiler instead of runtime. I probably need to find a better way to communicate to the runtime from the compiler.