Starting to move runtime to a global reference instead
This step is still definitely not safe. But as I make all the ways you interact with the runtime thread-safe, It should be fine to have this global access here instead of passing things around. Of course, making things global means I'm cutting off opportunity to have multiple. But I don't think I ever want that. Adding reset is interesting, because I should eventually have the ability to go back to a point in time. So snapshotting the runtime and recording changes to it should be something I support in the long run.