Very minor refactor from claude
It's funny. I have gotten out of the habit of simple changes like this. In some ways, it reads so much nicer. In other, I have found some value in all logic being colocated. It will be interesting to see how/if coding styles change with the use of AI.