Binary trees is faster now
We were doing a heavy weight atomic for gc pause. We don't need to do that. In general, I need to make the gc stuff better. But I don't see a reason that reading needs to be atomic. We will just catch it later.

There was a hang on ci for the last commit. Doubt this fixes anything. Will investigate if it hangs again.